% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogCountry.R
\name{demogCountry}
\alias{demogCountry}
\title{Project population scenarios across countries.}
\usage{
demogCountry(popc = NA, rlc = NA)
}
\arguments{
\item{popc}{File (RDS) with population parameters for differnt scenarios.}

\item{rlc}{River lengths for each scenario. From "res.Tab.R".}
}
\value{
Creates dataframe with 50 year population projections.
}
\description{
Generates dataframe with population projections.
}
\details{
Title
}
\examples{
\dontrun{

lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
popc <- "C:\\\\Users\\\\Darren\\\\Documents\\\\2018 Unifilis demography\\\\analysis\\\\dfpop.RDS"
demogC <- demogCountry(popc = popc, rlc = lt$rlc)

}
}
