% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demogCountryBasin.R
\name{demogCountryBasin}
\alias{demogCountryBasin}
\title{Project population scenarios across countries and basins.}
\usage{
demogCountryBasin(popc = NA, rlcb = NA)
}
\arguments{
\item{popc}{File (RDS) with population parameters for differnt scenarios.}

\item{rlcb}{River lengths for each scenario, from "resTab.R".}
}
\value{
Creates dataframe with 50 year population projections.
}
\description{
Generates dataframe with population projections.
}
\details{
Title
}
\examples{
\dontrun{

lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
popc <- "C:\\\\Users\\\\Darren\\\\Documents\\\\2018 Unifilis demography\\\\analysis\\\\dfpop.RDS"
demogC <- demogCountryBasin(popc = popc, rlcb = lbt$rlcb)

}
}
