% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopParam.R
\name{PopParam}
\alias{PopParam}
\title{Create data.frame with demographic parameters}
\usage{
PopParam(species = "Podocnemis unifilis", make_rds = FALSE)
}
\arguments{
\item{species}{Name of species. Default is "Podocnemis unifilis"}

\item{make_rds}{Logical (TRUE/FALSE). Should .RDS format be written.
Default is FALSE}
}
\value{
Creates data.frame with parameters used in "PopProj.R".
}
\description{
Creates data.frame with parameters for use in subsequent
population projection function ("PopProj.R")
}
\examples{
\dontrun{
dfpop <- PopParam(species = "Podocnemis unifilis", make_rds = FALSE)
}

}
