% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopPrep.R
\name{PopPrep}
\alias{PopPrep}
\title{Prepare data for population projections.}
\usage{
PopPrep(x = NA, riverl = NA)
}
\arguments{
\item{x}{Data.frame with population parameters. Created by "PopParam.R".}

\item{riverl}{Data.frame with river lengths per category of interest. 
Created by "resTab.R".}
}
\value{
List with population parameter matrix and river lengths.
}
\description{
Creates list holding parameter matrix and river lengths.
}
\examples{
\dontrun{
#1) River length coverage
# Run, 4 - 5 hours depending on computer and memory allocation
lsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
          pBasinC = BC, riv = rin, rastAc = ras1, make_shape = FALSE)

# Folder with shapefiles with 1km rvier sections
rp <- system.file("shape/shapes_rivers3395", package="cmartr")
lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
atest <- lt$rlcb
# add key and make sure all levels present
atest$namekey <- paste(atest$BASIN_N, atest$name, atest$subbasn, sep = "_")
riverl <- expand.grid(namekey = unique(atest$namekey), 
                     accessible = unique(atest$accessible) )
riverl <- merge(riverl, atest, all.x=TRUE)
selNA <- which(is.na(riverl$tot_km))
coln <- c("tot_km", "tot_notPA", "tot_PA", "tot_Ind", "tot_SP", "tot_use")
riverl[selNA, coln] <- 0

#2) Data.frame with population parameters
dfpop <- PopParam(species = "Podocnemis unifilis", make_rds = FALSE)

#3) Create list with population parameters for different scenarios
# Run. Few seconds 33.5 Mb
# Create list with population parameters for different scenarios
# across river lengths per geographic/political coverage class: 
# basin, country, subbasin, accessible, protected area...
l.gpop <- plyr::dlply(dfpop, c("akey"), PopPrep, riverl=riverl)
}
}
