% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopProj.R
\name{PopProj}
\alias{PopProj}
\title{Project populations.}
\usage{
PopProj(x, write_csv = FALSE, write_db = FALSE)
}
\arguments{
\item{x}{List with population parameters for different scenarios.
Creates using "PopPrep.R"}

\item{write_csv}{Logical (TRUE/FALSE). Should csv tables be written.}

\item{write_db}{Logical (TRUE/FALSE). Should postgre tables be written.}
}
\value{
Results as .csv and/or postgresql tables. A lookup table is returned for users that 
is then used for producing reaults.
}
\description{
Projects populations through different scenarios across 
species range.
}
\details{
Title
}
\examples{
\dontrun{
#1) River length coverage
lsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
pBasinC = BC, riv = rin, rastAc = ras1, make_shape = FALSE)
rp <- system.file("shape/shapes_rivers3395", package="cmartr")
lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
atest <- lt$rlcb
# add key and make sure all levels are represented
atest$namekey <- paste(atest$BASIN_N, atest$name, atest$subbasn, sep = "_")
riverl <- expand.grid(namekey = unique(atest$namekey), 
                     accessible = unique(atest$accessible) )
riverl <- merge(riverl, atest, all.x=TRUE)
selNA <- which(is.na(riverl$tot_km))
riverl[selNA, c("tot_km", "tot_notPA", "tot_PA", "tot_Ind", "tot_SP", "tot_use")] <- 0

#2) Data frame with population parameters created from "PopParam.R" 
dfpop <- readRDS("inst/other/dfpop.RDS") 

#3) Create list with population parameters for different scenarios
# across river lengths per geographic/political coverage class: 
# basin, country, subbasin, accessible, protected area...
l.gpop <- plyr::dlply(dfpop, c("akey"), PopPrep, riverl=riverl)

#4) Project scenarios across species range
# takes 5 hours and writes 10 GB of results.
# Do not run unless you really want to.....
dflup <- plyr::ldply(l.gpop, PopProj, write_csv = TRUE, 
write_db = FALSE)
}
}
