% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopScenAcess.R
\name{PopScenAcess}
\alias{PopScenAcess}
\title{Extract and combine data for different scenarios.}
\usage{
PopScenAcess(x)
}
\arguments{
\item{x}{Lookup dataframe created by "PopProj.R".}
}
\value{
List holding three dataframes, with results for each scenario.
}
\description{
Obtains results from "PopProj.R" and combines to 
enable subsequent processing. Adapted to evaluate sensitivity to
differences in acessibility.
Ensures population values do not inflate beyond initial values (max 100 *).
}
\details{
Title

Obtains results from "PopProj.R" and combines to 
enable subsequent processing. 
Ensures population values do not inflate beyond initial values (max 100 *).
}
\examples{
\dontrun{
#1) River length coverage
lsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
pBasinC = BC, riv = rin, rastAc = ras1, make_shape = FALSE)
rp <- system.file("shape/shapes_rivers3395", package="cmartr")
lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
atest <- lt$rlcb
# add key and make sure all levels are represented
atest$namekey <- paste(atest$BASIN_N, atest$name, atest$subbasn, sep = "_")
riverl <- expand.grid(namekey = unique(atest$namekey), 
                     accessible = unique(atest$accessible) )
riverl <- merge(riverl, atest, all.x=TRUE)
selNA <- which(is.na(riverl$tot_km))
riverl[selNA, c("tot_km", "tot_notPA", "tot_PA", "tot_Ind", "tot_SP", "tot_use")] <- 0

#2) Data frame with population parameters created from "PopParam.R" 
dfpop <- readRDS("inst/other/dfpop.RDS") 

#3) Create list with population parameters for different scenarios
# across river lengths per geographic/political coverage class: 
# basin, country, subbasin, accessible, protected area...
l.gpop <- plyr::dlply(dfpop, c("akey"), PopPrep, riverl=riverl)

#4) Project scenarios across species range
# takes 5 hours and writes 10 GB of results.
# Do not run unless you really want to.....
dflup <- plyr::ldply(l.gpop, PopProj, write_csv = TRUE, 
write_db = FALSE)

#5) Get results from 3 scenarios
lscen <- PopScen(dflup)
}
}
