% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepBasin.R
\name{prepBasin}
\alias{prepBasin}
\title{Prepare basin polygons}
\usage{
prepBasin(Bain = NA, make_shape = FALSE)
}
\arguments{
\item{Bain}{Input file with basin polygons}

\item{make_shape}{Logical (TRUE/FALSE). Generate polygon shapefile 
covering basins with species locations ("SpeciesBasin.shp").}
}
\value{
Files for subsequent use in "PrepPAcover.R".
}
\description{
Generates files for subsequent use. 
Optionally creates "SpeciesBasin.shp" .
}
\details{
Takes georeferenced polygons of river basins
}
\examples{
\dontrun{
# These large files are not available via github package. 
# Download "shapes.zip" from :
# https://drive.google.com/open?id=1QQArA7pPLemUVQTKx7PxigsQXKpOG0YQ
# For the code below to work, 
# extract files in "shapes.zip" to cmartr/inst/shape 
B <- system.file("shape/amazon_orinoco.shp", package="cmartr")
sfclean <- cmartr::prepBasin(Bain = pBasin)
}
}
