% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepPAcountry.R
\name{prepPAcountry}
\alias{prepPAcountry}
\title{Prepare polygons for PA calculations}
\usage{
prepPAcountry(PAcov = NA, pBasin = NA, make_shape = FALSE)
}
\arguments{
\item{PAcov}{Input file with PA cover polygons. Created from prepPAcover.}

\item{pBasin}{Input file with basin polygons.}

\item{make_shape}{Logical (TRUE/FALSE). Generate polygon shapefile of
protected areas in 3 classes in species range. With country and basin attributes.}
}
\value{
Files for future use. Optionally exports
polygon as shapefile "basinpacover.shp".
}
\description{
Generates files for subsequent use. Optionally exports
polygon as shapefile "basinpacover.shp".
}
\examples{
\dontrun{
# Load shapefiles needed
# These large files are not available via github package. 
# Download "shapes.zip" from :
# https://drive.google.com/open?id=1QQArA7pPLemUVQTKx7PxigsQXKpOG0YQ
# For the code below to work, 
# extract files in "shapes.zip" to cmartr/inst/shape .
Pcov <- system.file("shape/pacover.shp", package="cmartr")
B <- system.file("shape/amazon_orinoco.shp", package="cmartr")
pacovc <- prepPAcountry(PAcov = Pcov, pBasin = B)
}
}
