% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTabcover.R
\name{prepTabcover}
\alias{prepTabcover}
\title{Prepare polygons for Table calculations}
\usage{
prepTabcover(pBasin = NA, pBasinSp = NA, pBasinC = NA, riv = NA,
  make_shape = FALSE)
}
\arguments{
\item{pBasin}{Subbasin cover polygons.}

\item{pBasinSp}{Extent polygon polygon covering 
subbasins in species range of occurance.}

\item{pBasinC}{PA cover in subbasins with country.}

\item{riv}{River lines.}

\item{make_shape}{Logical (TRUE/FALSE). Generate polygon shapefiles of
......}
}
\value{
List with objects of class sf with attribute data 
used to generate result tables. Optionally writes shapefiles.
}
\description{
Generates sf objects with attributes for Table values. Optionally exports
polygon as shapefile.
}
\examples{
\dontrun{
# Load shapefiles needed
# These large files are not available via github package. 
# Download "shapes.zip" from :
# https://drive.google.com/open?id=1QQArA7pPLemUVQTKx7PxigsQXKpOG0YQ
# For the code below to work, 
# extract files in "shapes.zip" to cmartr/inst/shape .
# This function needs 4 shapefiles as listed below.

# Subbasin cover polygons
B <- system.file("shape/amazon_orinoco.shp", package="cmartr")
# Extent polygon covering subbasins in species range of occurance.
# From prepBasin.R .
Bsp <- system.file("shape/speciesBasin.shp", package="cmartr")
# PA cover in subbasins with country. From prepPAcountry.R 
BC <- system.file("shape/basinpacover.shp", package="cmartr")
# River lines.
rin <- system.file("shape/sa_strwgs84.shp", package="cmartr")

# run
listsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
  pBasinC = BC, riv = rin, make_shape = FALSE)
}
}
