% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resFig2.R
\name{resFig2}
\alias{resFig2}
\title{Create Figure 2}
\usage{
resFig2(listsf = NA, listTab = NA, pBasin = NA, pBasinSp = NA,
  pBasinAcc = NA, rastAc = NA, make_png = FALSE)
}
\arguments{
\item{listsf}{List of five sf objects created by prepTabcover.R}

\item{listTab}{List of data.frames created by resTab.R}

\item{pBasin}{Subbasin cover polygons.}

\item{pBasinSp}{Extent polygon covering 
subbasins in species range of occurance.}

\item{pBasinAcc}{Subbasin cover polygons with human access proportion,
from demog_river.R (SAGA (grid tools, grids stats for poly)).}

\item{rastAc}{Acessibility raster.}

\item{make_png}{Logical (TRUE/FALSE). Generate png version of Fig 1.}
}
\value{
List of plots created by ggplot2. Optionally exports as png.
}
\description{
Generates Figure 2. Returns list with ggplot2 figures. 
Optionally exports as png.
}
\examples{
\dontrun{
# Load shapefiles needed
# These large files are not available via github package. 
# Download "shapes.zip" from :
# https://drive.google.com/open?id=1QQArA7pPLemUVQTKx7PxigsQXKpOG0YQ
# For the code below to work, 
# extract files in "shapes.zip" to cmartr/inst/shape .
# This function needs 3 shapefiles as listed below.

# Subbasin cover polygons
B <- system.file("shape/amazon_orinoco.shp", package="cmartr")
# Extent polygon covering subbasins in species range of occurance.
# From prepBasin.R .
Bsp <- system.file("shape/speciesBasin.shp", package="cmartr")
# Human access from demog_river.R . Need to add earlier in prep stages.
BAc <- system.file("shape/basinsum4326.shp", package="cmartr")
# Human aceesibility raster
ras1 <- system.file("raster/uas.grd", package="cmartr")
# run
lsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
  pBasinC = BC, riv = rin, make_shape = FALSE)
lt <- resTab(listsf = lsf)
lfig2 <- resFig2(listsf = lsf, listTab = lt, pBasin = B, 
  pBasinSp = Bsp, pBasinAcc = BAc, rastAc = ras1, make_png = FALSE)

}
}
