% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resFig3.R
\name{resFig3}
\alias{resFig3}
\title{Create and export Figure 3.}
\usage{
resFig3(x)
}
\arguments{
\item{x}{Data.frame from cmartr::PopParam}
}
\value{
pdf of Figure 3.
}
\description{
Generates Figure 3.
}
\details{
Title
}
\examples{
\dontrun{
dfpop <- cmartr::PopParam(species = "Podocnemis unifilis", 
make_rds = FALSE)
resFig3(dfpop)
}
}
