% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resFig4.R
\name{resFig4}
\alias{resFig4}
\title{Create and export Figure 4.}
\usage{
resFig4(x)
}
\arguments{
\item{x}{Data.frame created by "proj.rivl.R".}
}
\value{
Exports Figure 4 as .pdf file.
}
\description{
Generates Figure 4. Needs updating to take results from 
"PopProj.R" and "PopScen.R"
}
\examples{
\dontrun{
dfres <- 
readRDS(
"C:\\\\Users\\\\Darren\\\\Documents\\\\ms\\\\unpublished\\\\2018 Unifilis demography\\\\analises\\\\dfpopres.RDS"
)
resFig4(dfres)
}
}
