% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTab.R
\name{resTab}
\alias{resTab}
\title{Create Table 1 and Table 2}
\usage{
resTab(listsf = NA)
}
\arguments{
\item{listsf}{List of five sf objects created by prepTabcover.R}
}
\value{
Creates html tables with values used in article.
}
\description{
Generates html files with Table values.
}
\examples{
\dontrun{
lsf <-  prepTabcover(pBasin = B, pBasinSp = Bsp, 
pBasinC = BC, riv = rin, make_shape = FALSE)
resTab <- (listsf = lsf)
}
}
