% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTabDemog.R
\name{resTabDemog}
\alias{resTabDemog}
\title{Create summary for population scenarios in Table 1.}
\usage{
resTabDemog(demogC = NA)
}
\arguments{
\item{demogC}{Country level population projections from "demogCountry.R"}
}
\value{
Data.frame with country level summary of population projections for three scenarios.
}
\description{
Used in Table 1.
Generates dataframe with country level summary of population projections.
}
\details{
Title
}
\examples{
\dontrun{

lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
popc <- "C:\\\\Users\\\\Darren\\\\Documents\\\\2018 Unifilis demography\\\\analysis\\\\dfpop.RDS"
demogC <- demogCountry(popc = popc, rlc = lt$rlc)
dfsum <- resTabDemog(demogC = demogC)
}
}
