% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTabDemogBasin.R
\name{resTabDemogBasin}
\alias{resTabDemogBasin}
\title{Create summary for population scenarios in Table 1.}
\usage{
resTabDemogBasin(demogCB = NA, make_html = FALSE)
}
\arguments{
\item{demogCB}{Country and basin level population projections from "demogCountryBasin.R"}

\item{make_html}{Logical (TRUE/FALSE). Should html tables be written.}
}
\value{
Data.frame with country level summary of population projections for three scenarios.
}
\description{
Used in Table 1.
Generates dataframe with country level summary of population projections.
}
\details{
Title
}
\examples{
\dontrun{

lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
popc <- "C:\\\\Users\\\\Darren\\\\Documents\\\\2018 Unifilis demography\\\\analysis\\\\dfpop.RDS"
demogCB <- demogCountryBasin(popc = popc, rlcb = lt$rlc)
dfsum <- resTabDemogBasin(demogCB = demogC, make_html = FALSE)
}
}
