% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepBasin.R
\name{prepBasin}
\alias{prepBasin}
\title{Prepare basin polygons}
\usage{
prepBasin(Bain = NA)
}
\arguments{
\item{Bain}{Input file with basin polygons}
}
\value{
Files for subsequent use.
}
\description{
Generates files for subsequent use.
}
\details{
Takes georeferenced polygons of river basins
}
\examples{
\dontrun{
}
}
