% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepPA.R
\name{prepPA}
\alias{prepPA}
\title{Prepare protected area polygons}
\usage{
prepPA(PAin = NA)
}
\arguments{
\item{PAin}{File with protected area polygons}
}
\value{
Generates files for subsequent use.
}
\examples{
\dontrun{
}
}
