% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepPA.R
\name{prepPA}
\alias{prepPA}
\title{Prepare protected area polygons}
\usage{
prepPA(PAin = NA)
}
\arguments{
\item{PAin}{Input file with protected area polygons}
}
\value{
Files for subsequent use.
}
\description{
Generates files for subsequent use in "PrepPAcover.R".
}
\details{
Takes georeferenced polygons of protected areas from the World Database on Protected Areas (WDPA, downloaded from https://protectedplanet.net/ on 6 December 2017).
Includes subset of terrestrial South American protected areas that are in the raster grid.
I.e. touching, intersecting, within bounding box created from 150 km  buffer around the catchments. 
 Cleans and identifies 3 classes: Indigenous lands, Strict Protection and Use.
 Definitions from https://www.iucn.org/theme/protected-areas/about/protected-area-categories .
}
\examples{
\dontrun{
# Load shapefiles needed
# These large files are not available via github package. 
# Download "shapes.zip" from :
# https://drive.google.com/open?id=1QQArA7pPLemUVQTKx7PxigsQXKpOG0YQ
# For the code below to work, 
# extract files in "shapes.zip" to cmartr/inst/shape .
PA <- system.file("shape/wdpaselect.shp", package="cmartr")
sf.pa <- cmartr::prepPA(PAin = PA)
}
}
