% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepPAcover.R
\name{prepPAcover}
\alias{prepPAcover}
\title{Prepare polygons for PAcover calculations}
\usage{
prepPAcover(pPA = NA, pBasin = NA, pBasinSp = NA)
}
\arguments{
\item{pPA}{Input file with protected area polygons.}

\item{pBasin}{Input file with basin polygons.}

\item{pBasinSp}{Input file with polygon covering basins in species range of occurance.}
}
\value{
Polygons of area covered by protected areas.
}
\description{
Generates files for subsequent use.
}
\details{
To remove overlapping areas with different PA classifications.
These are big areas and processing takes time: 3 - 4 hours on a laptop.
}
\examples{
\dontrun{
# load shapefiles needed
Bsp <- system.file("shape/speciesBasin.shp", package="cmartr")
PA <- system.file("shape/wdpaselect.shp", package="cmartr")
B <- system.file("shape/amazon_orinoco.shp", package="cmartr")
#add id for dissolved basin polygon
sfclean$sbid <- "P.unifilis"
pacover <- prepPAcover(pPA = PA, pBasin = B, pBasinSp = Bsp)
}
}
