% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resFig3.R
\name{resFig3}
\alias{resFig3}
\title{Create and export Figure 3.}
\usage{
resFig3(x)
}
\arguments{
\item{x}{Data.frame created by "proj.rivl.R".}
}
\value{
Exports Figure 3 as .pdf file.
}
\description{
Generates Figure 3. Needs updating to take results from 
"PopProj.R" and "PopScen.R"
}
\examples{
\dontrun{
dfres <- 
readRDS(
"C:\\\\Users\\\\Darren\\\\Documents\\\\ms\\\\unpublished\\\\2018 Unifilis demography\\\\analises\\\\dfpopres.RDS"
)
resFig3(dfres)
}
}
