% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resFig5.R
\name{resFig5}
\alias{resFig5}
\title{Create and export Figure 5.}
\usage{
resFig5(x)
}
\arguments{
\item{x}{Data.frame created by "proj.rivl.R".}
}
\value{
Exports Figure 5 as .png file.
}
\description{
Generates Figure 5. Needs updating to take results from 
"PopProj.R" and "PopScen.R"
}
\examples{
\dontrun{
dfpop.res <- 
readRDS(
"C:\\\\Users\\\\Darren\\\\Documents\\\\ms\\\\unpublished\\\\2018 Unifilis demography\\\\analises\\\\dfpopres.RDS"
)
resFig5(dfres)
}
}
