% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resTabDemog.R
\name{resTabDemog}
\alias{resTabDemog}
\title{Create summary for population scenarios in Table 1.}
\usage{
resTabDemog(ldemog = NA, make_html = FALSE)
}
\arguments{
\item{ldemog}{List with population projections from "PopScen.R"}

\item{make_html}{Logical (TRUE/FALSE). Should html tables be written.}
}
\value{
Data.frame with summaries of population projections for three scenarios.
}
\description{
Used in Table 1, Supplemental Tables.
Generates dataframe with summaries of population projections.
}
\details{
Title
}
\examples{
\dontrun{
#1) River length coverage
lsf <- prepTabcover(pBasin = B, pBasinSp = Bsp, 
pBasinC = BC, riv = rin, rastAc = ras1, make_shape = FALSE)
rp <- system.file("shape/shapes_rivers3395", package="cmartr")
lt <- resTab(listsf = lsf, input_rp = rp, make_html = FALSE)
atest <- lt$rlcb
# add key and make sure all levels are represented
atest$namekey <- paste(atest$BASIN_N, atest$name, atest$subbasn, sep = "_")
riverl <- expand.grid(namekey = unique(atest$namekey), 
                     accessible = unique(atest$accessible) )
riverl <- merge(riverl, atest, all.x=TRUE)
selNA <- which(is.na(riverl$tot_km))
coln <- c("tot_km", "tot_notPA", "tot_PA", "tot_Ind", "tot_SP", "tot_use")
riverl[selNA, coln)] <- 0

#2) Data frame with population parameters created from "PopParam.R" 
dfpop <- readRDS("inst/other/dfpop.RDS") 

#3) Create list with population parameters for different scenarios
# across river lengths per geographic/political coverage class: 
# basin, country, subbasin, accessible, protected area...
l.gpop <- plyr::dlply(dfpop, c("akey"), PopPrep, riverl=riverl)

#4) Project scenarios across species range
# takes 5 hours and writes 10 GB of results.
# Do not run unless you really want to.....
dflup <- plyr::ldply(l.gpop, PopProj, write_csv = TRUE, 
write_db = FALSE)

#5) Get results from 3 scenarios
lscen <- PopScen(dflup)

#6) Produce summary tables.
dfsum <- resTabDemog(ldemog = lscen, make_html = FALSE)
}
}
